Create file name moved to the FileNameStore
New test project created for FilePersistence. Naming the project as 'CLAL*' as file length reached. Will rename all other projects into this format.
moved the tests that was checking the file name to the new test project created for the persistence layer. Keeping this as '.Test' as we might have some integration tests too here.